Create GradeChecker.py
Part of the CFG intro to coding challenge, this was challenge 1. Create code that would take an inputted grade, check whether the next multiple of 5 was less than 3 marks away if it was you upgrade the grade to the new mark. If a grade was lower than 40 it was a fail, between 40 and 80 was a pass and over 80 was a distinction.